refactor: employee details form reusability
